更新Buffer相关实现

+++ source/renderer/vulkan/RenderPassVk.cpp
@@ -610,8 +610,8 @@ namespace Nix {
 		m_viewport.x = m_viewport.y = 0;
 		m_viewport.minDepth = 0.0f;
 		m_viewport.maxDepth = 1.0f;
-		m_viewport.width = _width;
-		m_viewport.height = _height;
+		m_viewport.width = (float)_width;
+		m_viewport.height = (float)_height;
 		m_scissor.extent.width = _width;
 		m_scissor.extent.height = _height;
 		m_scissor.offset = { 0, 0 };
